+++ CHANGELOG.md
@@ -41,6 +41,7 @@ The format is based on Keep a Changelog, and this project adheres to Semantic Ve
 - Simplified `ExitedValidatorIndices`.
 - Simplified `EjectedValidatorIndices`.
 - `engine_newPayloadV4`,`engine_getPayloadV4` are changes due to new execution request serialization decisions, [PR](https://github.com/prysmaticlabs/prysm/pull/14580)
+- Fixed various small things in state-native code.
 - Use ROBlock earlier in block syncing pipeline. 
 - Changed the signature of `ProcessPayload`.
 - Only Build the Protobuf state once during serialization.